move AttributeGuard into iocontainers
